Passes lint. Sauce needs testing, travis integration.
