+++ TODO.md
@@ -3,13 +3,11 @@ Tasks
 
 ## POC Branches
 
-* Add gulp AMD tests. [feature-karma-amd](https://github.com/FormidableLabs/chai-jq/compare/feature-karma-amd)
-* Add blanket to HTML tests. [chore-blanket](https://github.com/FormidableLabs/chai-jq/compare/chore-blanket)
+* Add karma AMD tests. [feature-karma-amd](https://github.com/FormidableLabs/chai-jq/compare/feature-karma-amd)
 
 ## Next Up
 
 * Upgrade all NPM libs.
-* Bower cleanup and upgrades.
 
 * GFM Headings/TOC: Switch pre-processing in grunt to use GFM headings / anchors
   and add extra output in HTML (?).
